work with work double

+++ app/src/main/java/com/example/tracker/data/SavedWalk.kt
@@ -1,3 +1,3 @@
 package com.example.tracker.data
 
-class SavedWalk(val location: String, val distance: Int)
+class SavedWalk(val location: String, val distance: Double)
